requirements.txt
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==36.0.2
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.2.6
opencv-python==4.11.0.86
pillow==11.2.1
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-engineio==4.12.1
python-socketio==5.13.0
python3-xlib==0.15
pytweening==1.2.0
simple-websocket==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.14.0
Werkzeug==3.1.3
wsproto==1.2.0
